test: migrates quality gates to ts and segregates class responsibility
